--- printer/src/main/java/com/dantsu/escposprinter/barcode/Barcode39.java
@@ -1,21 +1,0 @@
-package com.dantsu.escposprinter.barcode;
-
-import com.dantsu.escposprinter.EscPosPrinterSize;
-import com.dantsu.escposprinter.EscPosPrinterCommands;
-import com.dantsu.escposprinter.exceptions.EscPosBarcodeException;
-
-public class Barcode39 extends Barcode {
-    public Barcode39(EscPosPrinterSize printerSize, String code, float widthMM, float heightMM, int textPosition) throws EscPosBarcodeException {
-        super(printerSize, EscPosPrinterCommands.BARCODE_TYPE_39, code, widthMM, heightMM, textPosition);
-    }
-
-    @Override
-    public int getCodeLength() {
-        return this.code.length();
-    }
-
-    @Override
-    public int getColsCount() {
-        return (this.getCodeLength() + 4) * 16;
-    }
-}
