--- printer/src/main/java/com/dantsu/escposprinter/exceptions/EscPosBarcodeException.java
@@ -1,7 +1,0 @@
-package com.dantsu.escposprinter.exceptions;
-
-public class EscPosBarcodeException extends Exception {
-    public EscPosBarcodeException(String errorMessage) {
-        super(errorMessage);
-    }
-}
