--- printer/src/main/java/com/dantsu/escposprinter/exceptions/EscPosConnectionException.java
@@ -1,7 +1,0 @@
-package com.dantsu.escposprinter.exceptions;
-
-public class EscPosConnectionException extends Exception {
-    public EscPosConnectionException(String errorMessage) {
-        super(errorMessage);
-    }
-}
