--- printer/src/main/java/com/dantsu/escposprinter/exceptions/EscPosEncodingException.java
@@ -1,7 +1,0 @@
-package com.dantsu.escposprinter.exceptions;
-
-public class EscPosEncodingException extends Exception {
-    public EscPosEncodingException(String errorMessage) {
-        super(errorMessage);
-    }
-}
