--- printer/src/main/java/com/dantsu/escposprinter/exceptions/EscPosParserException.java
@@ -1,7 +1,0 @@
-package com.dantsu.escposprinter.exceptions;
-
-public class EscPosParserException extends Exception {
-    public EscPosParserException(String errorMessage) {
-        super(errorMessage);
-    }
-}
