+++ printer/src/main/java/eloom/holybean/escpos/barcode/Barcode128.java
@@ -1,0 +1,21 @@
+package eloom.holybean.escpos.barcode;
+
+import eloom.holybean.escpos.EscPosPrinterSize;
+import eloom.holybean.escpos.EscPosPrinterCommands;
+import eloom.holybean.escpos.exceptions.EscPosBarcodeException;
+
+public class Barcode128 extends Barcode {
+    public Barcode128(EscPosPrinterSize printerSize, String code, float widthMM, float heightMM, int textPosition) throws EscPosBarcodeException {
+        super(printerSize, EscPosPrinterCommands.BARCODE_TYPE_128, code, widthMM, heightMM, textPosition);
+    }
+
+    @Override
+    public int getCodeLength() {
+        return this.code.length();
+    }
+
+    @Override
+    public int getColsCount() {
+        return (this.getCodeLength() + 5) * 11;
+    }
+}
