+++ printer/src/main/java/eloom/holybean/escpos/barcode/Barcode39.java
@@ -1,0 +1,21 @@
+package eloom.holybean.escpos.barcode;
+
+import eloom.holybean.escpos.EscPosPrinterSize;
+import eloom.holybean.escpos.EscPosPrinterCommands;
+import eloom.holybean.escpos.exceptions.EscPosBarcodeException;
+
+public class Barcode39 extends Barcode {
+    public Barcode39(EscPosPrinterSize printerSize, String code, float widthMM, float heightMM, int textPosition) throws EscPosBarcodeException {
+        super(printerSize, EscPosPrinterCommands.BARCODE_TYPE_39, code, widthMM, heightMM, textPosition);
+    }
+
+    @Override
+    public int getCodeLength() {
+        return this.code.length();
+    }
+
+    @Override
+    public int getColsCount() {
+        return (this.getCodeLength() + 4) * 16;
+    }
+}
