+++ printer/src/main/java/eloom/holybean/escpos/barcode/BarcodeEAN13.java
@@ -1,0 +1,17 @@
+package eloom.holybean.escpos.barcode;
+
+import eloom.holybean.escpos.EscPosPrinterCommands;
+import eloom.holybean.escpos.EscPosPrinterSize;
+import eloom.holybean.escpos.exceptions.EscPosBarcodeException;
+
+public class BarcodeEAN13 extends BarcodeNumber {
+
+    public BarcodeEAN13(EscPosPrinterSize printerSize, String code, float widthMM, float heightMM, int textPosition) throws EscPosBarcodeException {
+        super(printerSize, EscPosPrinterCommands.BARCODE_TYPE_EAN13, code, widthMM, heightMM, textPosition);
+    }
+
+    @Override
+    public int getCodeLength() {
+        return 13;
+    }
+}
