+++ printer/src/main/java/eloom/holybean/escpos/exceptions/EscPosBarcodeException.java
@@ -1,0 +1,7 @@
+package eloom.holybean.escpos.exceptions;
+
+public class EscPosBarcodeException extends Exception {
+    public EscPosBarcodeException(String errorMessage) {
+        super(errorMessage);
+    }
+}
