+++ printer/src/main/java/eloom/holybean/escpos/exceptions/EscPosConnectionException.java
@@ -1,0 +1,7 @@
+package eloom.holybean.escpos.exceptions;
+
+public class EscPosConnectionException extends Exception {
+    public EscPosConnectionException(String errorMessage) {
+        super(errorMessage);
+    }
+}
