+++ printer/src/main/java/eloom/holybean/escpos/exceptions/EscPosEncodingException.java
@@ -1,0 +1,7 @@
+package eloom.holybean.escpos.exceptions;
+
+public class EscPosEncodingException extends Exception {
+    public EscPosEncodingException(String errorMessage) {
+        super(errorMessage);
+    }
+}
