+++ printer/src/main/java/eloom/holybean/escpos/exceptions/EscPosParserException.java
@@ -1,0 +1,7 @@
+package eloom.holybean.escpos.exceptions;
+
+public class EscPosParserException extends Exception {
+    public EscPosParserException(String errorMessage) {
+        super(errorMessage);
+    }
+}
