+++ printer/src/main/java/eloom/holybean/escpos/textparser/IPrinterTextParserElement.java
@@ -1,0 +1,10 @@
+package eloom.holybean.escpos.textparser;
+
+import eloom.holybean.escpos.EscPosPrinterCommands;
+import eloom.holybean.escpos.exceptions.EscPosConnectionException;
+import eloom.holybean.escpos.exceptions.EscPosEncodingException;
+
+public interface IPrinterTextParserElement {
+    int length() throws EscPosEncodingException;
+    IPrinterTextParserElement print(EscPosPrinterCommands printerSocket) throws EscPosEncodingException, EscPosConnectionException;
+}
